Adding slider components
